requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
json-repair>=0.30.0